# src/docs/combined-swagger.yaml
openapi: 3.0.0
info:
  title: Synthetix Cache API
  version: 1.0.0
  description: API documentation for the Synthetix Cache API
servers:
  - url: http://localhost:3001
  - url: https://synthetix-cache-api-f7db01015ec1.herokuapp.com
paths:
...
  /core-account-delegations/staker-count/{chain}:
    get:
      summary: Get the unique staker count
      tags:
        - Core Account Delegations
      parameters:
        - in: path
          name: chain
          schema:
            type: string
          required: false
          description: The chain to get the staker count for
      responses:
        200:
          description: The unique staker count
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                  example: '305'
        500:
          description: Server error
  /core-account-delegations/cumulative-unique-stakers/{chain}:
    get:
      summary: Get cumulative unique staker data
      tags: 
        - Core Account Delegations
      parameters:
        - in: path
          name: chain
          schema:
            type: string
          required: false
          description: The chain to get the cumulative unique staker data for
      responses:
        200:
          description: The cumulative unique staker data
          content:
            application/json:
              schema:
                type: object
                properties:
                  chain:
                    type: string
                    example: base
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        ts:
                          type: string
                          format: date-time
                          example: '2024-07-14T09:00:00Z'
                        cumulative_staker_count:
                          type: integer
                          example: 150
                        pool_id:
                          type: integer
                          example: 1
                        collateral_type:
                          type: string
                          example: '0xC74eA762cF06c9151cE074E6a569a5945b6302E7'
        404:
          description: Cumulative unique staker data not found
        500:
          description: Server error
  /core-account-delegations/all/{chain}:
    get:
      summary: Get all core account delegations data
      tags:
        - Core Account Delegations
      parameters:
        - in: path
          name: chain
          schema:
            type: string
          required: false
          description: The chain to get the data for
      responses:
        200:
          description: All core account delegations data
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 1450
                    created_at:
                      type: string
                      format: date-time
                      example: '2024-06-27T15:00:50.345Z'
                    updated_at:
                      type: string
                      format: date-time
                      example: '2024-06-27T15:00:50.345Z'
                    ts:
                      type: string
                      format: date-time
                      example: '2024-06-27T14:58:45.000Z'
                    chain:
                      type: string
                      example: arbitrum
                    account_id:
                      type: string
                      example: '170141183460469231731687303715884105785'
                    pool_id:
                      type: integer
                      example: 1
                    collateral_type:
                      type: string
                      example: '0x912CE59144191C1204E64559FE8253a0e49E6548'
                    amount_delegated:
                      type: number
                      format: double
                      example: 19128.7385237578
        404:
          description: Core account delegations data not found
        500:
          description: Server error
  /core-account-delegations/account/{accountId}:
    get:
      summary: Get core account delegations data by account ID
      tags:
        - Core Account Delegations
      parameters:
        - in: path
          name: accountId
          schema:
            type: string
          required: true
          description: The account ID to get the data for
      responses:
        200:
          description: Core account delegations data for the given account ID
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 879
                    created_at:
                      type: string
                      format: date-time
                      example: '2024-06-25T15:45:11.350Z'
                    updated_at:
                      type: string
                      format: date-time
                      example: '2024-06-25T15:45:11.350Z'
                    ts:
                      type: string
                      format: date-time
                      example: '2023-12-15T08:04:45.000Z'
                    chain:
                      type: string
                      example: base
                    account_id:
                      type: string
                      example: '1'
                    pool_id:
                      type: integer
                      example: 1
                    collateral_type:
                      type: string
                      example: '0xC74eA762cF06c9151cE074E6a569a5945b6302E7'
                    amount_delegated:
                      type: number
                      format: double
                      example: 200
        404:
          description: Core account delegations data not found for this account ID
        500:
          description: Server error
  /core-account-delegations/ordered-by-account/{chain}:
    get:
      summary: Get core account delegations data ordered by account
      tags:
        - Core Account Delegations
      parameters:
        - in: path
          name: chain
          schema:
            type: string
          required: false
          description: The chain to get the data for
      responses:
        200:
          description: Core account delegations data ordered by account
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 879
                    created_at:
                      type: string
                      format: date-time
                      example: '2024-06-25T15:45:11.350Z'
                    updated_at:
                      type: string
                      format: date-time
                      example: '2024-06-25T15:45:11.350Z'
                    ts:
                      type: string
                      format: date-time
                      example: '2023-12-15T08:04:45.000Z'
                    chain:
                      type: string
                      example: base
                    account_id:
                      type: string
                      example: '1'
                    pool_id:
                      type: integer
                      example: 1
                    collateral_type:
                      type: string
                      example: '0xC74eA762cF06c9151cE074E6a569a5945b6302E7'
                    amount_delegated:
                      type: number
                      format: double
                      example: 200
        404:
          description: Core account delegations data not found
        500:
          description: Server error
  /core-account-delegations/cumulative-unique-stakers/summary/{chain}:
    get:
      summary: Get the cumulative unique stakers summary statistics
      tags:
        - Core Account Delegations
      parameters:
        - in: path
          name: chain
          schema:
            type: string
          required: true
          description: The chain to get the cumulative unique stakers summary statistics for
      responses:
        200:
          description: The cumulative unique stakers summary statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  current:
                    type: number
                    format: double
                    example: 305
                  delta_24h:
                    type: number
                    format: double
                    example: 10
                  delta_7d:
                    type: number
                    format: double
                    example: 25
                  delta_28d:
                    type: number
                    format: double
                    example: 50
                  delta_ytd:
                    type: number
                    format: double
                    example: 100
                  ath:
                    type: number
                    format: double
                    example: 310
                  atl:
                    type: number
                    format: double
                    example: 200
                  ath_percentage:
                    type: number
                    format: double
                    example: 98.39
                  atl_percentage:
                    type: number
                    format: double
                    example: 152.50
                  standard_deviation:
                    type: number
                    format: double
                    example: 15.5
        400:
          description: Bad request (invalid or missing chain parameter)
        500:
          description: Server error
  /core-account-delegations/daily-new-unique-stakers/{chain}:
    get:
      summary: Get daily new unique stakers data
      tags:
        - Core Account Delegations
      parameters:
        - in: path
          name: chain
          schema:
            type: string
          required: false
          description: The chain to get the daily new unique stakers data for. If not provided, returns data for all chains.
      responses:
        200:
          description: Daily new unique stakers data
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: object
                    properties:
                      ts:
                        type: string
                        format: date-time
                        example: "2023-05-01T00:00:00Z"
                      daily_new_unique_stakers:
                        type: integer
                        example: 42
        400:
          description: Bad request - Invalid chain parameter
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        404:
          description: Daily new unique stakers data not found
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /core-account-delegations/daily-new-unique-stakers/summary/{chain}:
    get:
      summary: Get summary statistics for daily new unique stakers
      tags:
        - Core Account Delegations
      parameters:
        - in: path
          name: chain
          schema:
            type: string
          required: true
          description: The chain to get the summary statistics for
      responses:
        200:
          description: Summary statistics for daily new unique stakers
          content:
            application/json:
              schema:
                type: object
                properties:
                  current:
                    type: integer
                    example: 42
                  delta_24h:
                    type: number
                    format: float
                    example: 5.5
                  delta_7d:
                    type: number
                    format: float
                    example: -2.3
                  delta_28d:
                    type: number
                    format: float
                    example: 10.1
                  delta_ytd:
                    type: number
                    format: float
                    example: 15.7
                  ath:
                    type: integer
                    example: 100
                  atl:
                    type: integer
                    example: 5
                  ath_percentage:
                    type: number
                    format: float
                    example: 42.0
                  atl_percentage:
                    type: number
                    format: float
                    example: 840.0
                  standard_deviation:
                    type: number
                    format: float
                    example: 12.3
        400:
          description: Bad request - Invalid or missing chain parameter
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        404:
          description: No data found for the specified chain
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        500:
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
...